refactor(checks): split package/file/service into module dirs
Decomposes the three per-check files (each >500 lines) into module
directories. Pure code-move: behavior byte-identical.

package/ (698 ->):
- apt.rs    query_apt_package_state + parse_apt_package_output (98)
- dnf.rs    query_dnf_package_state (84)
- mod.rs    check_package + execute/extract/validate + PackageParameters +
            PackageState(pub) + query_package_state + PackageManager +
            detect_package_manager + compare_package_state (265)
- tests.rs  15 unit tests (261)

file/ (547 ->):
- compare.rs  compare_file_state (83)
- mod.rs      check_file + execute/extract + FileParameters + FileState(pub) +
              query_file_state + parse_stat_output (245)
- tests.rs    12 unit tests (224)

service/ (583 ->):
- compare.rs  query_service_active_and_enabled_state + compare_service_state (127)
- mod.rs      check_service + execute/extract/validate + ServiceParameters +
              ServiceState(pub) + query_service_state (213)
- tests.rs    15 unit tests (253)

Visibility: PackageState/FileState/ServiceState promoted to pub struct (private
modules keep them effectively internal; shared with the compare/backend
submodules). Extracted helpers are pub(super) (visible to mod.rs + test module
within the subtree). checks/mod.rs + collect_check_functions registration
unchanged. Every file now <500 lines.
